requirements.txt
PyYAML
boto3
logging
